refactor: extract the secret cache lookup and value injection from the injector
The job push needs the same two mechanics the long poll already has: look
the secret references of a job up in the per-store caches, and replace the
placeholders in the job's variables with the values found. Both lived in
JobSecretInjector, wired into the batch bookkeeping of the long poll
(indexes in the batch, growth budget, dropping jobs), which the push has
no use for: it hands out one job at a time.

JobSecretValues now owns the two mechanics and the materialized values
behind them, and JobSecretInjector keeps the batch bookkeeping and
delegates. The values stay valid until reset(), which is what lets a
caller check the jobs of one activation together, so the class states that
contract for the second caller: a caller that checks a single job resets
before each check, or it could inject a value that has expired since.

No behaviour change. The nested Secret and SecretCheckResult records move
along with the mechanics that produce them.
