+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/job/JobSecretInjector.java
@@ -7,38 +7,28 @@
  */
 package io.camunda.zeebe.engine.processing.job;
 
-import static java.util.Comparator.comparingInt;
-
-import com.fasterxml.jackson.core.JsonPointer;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
 import io.camunda.secretstore.SecretStoreRegistry;
 import io.camunda.zeebe.engine.EngineConfiguration;
 import io.camunda.zeebe.engine.processing.deployment.model.element.SecretReference;
+import io.camunda.zeebe.engine.processing.job.JobSecretValues.Secret;
+import io.camunda.zeebe.engine.processing.job.JobSecretValues.SecretCheckResult;
 import io.camunda.zeebe.protocol.impl.record.value.job.JobBatchRecord;
 import io.camunda.zeebe.protocol.impl.record.value.job.JobRecord;
-import io.camunda.zeebe.protocol.impl.record.value.job.JobSecretReference;
 import io.camunda.zeebe.util.buffer.BufferUtil;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
-import org.agrona.DirectBuffer;
-import org.agrona.io.DirectBufferInputStream;
-import org.msgpack.jackson.dataformat.MessagePackFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * Injects cached secret values into the variables of activated jobs, in two steps before the
- * ACTIVATED event is appended:
+ * Injects cached secret values into the variables of the jobs activated by a long poll, in two
+ * steps before the ACTIVATED event is appended:
  *
  * <ol>
  *   <li>During batch collection, {@link #checkSecrets} looks up the secret references of every job
@@ -60,28 +50,18 @@ import org.slf4j.LoggerFactory;
  * reach the record that is written to the activation response and nowhere else. That keeps the
  * secret values on the response only: state, records, and logs keep the placeholders.
  *
- * <p>The replacement is textual: every occurrence of the placeholder in the addressed leaf is
- * replaced, including text that merely spells out the placeholder next to a real reference at the
- * same path. This can only surface a secret in a leaf that already receives that secret.
- *
- * <p>Only the long-poll activation response is covered. The job-push path (see {@code
- * BpmnJobActivationBehavior#publishWork}) is out of scope for this change and still streams the
- * unresolved placeholders.
+ * <p>The job-push path injects the same values into the job it streams, see {@code
+ * BpmnJobActivationBehavior#publishWork}.
  */
 public final class JobSecretInjector {
 
   private static final Logger LOGGER =
       LoggerFactory.getLogger(JobSecretInjector.class.getPackageName());
-  private static final String SECRET_REFERENCE_PREFIX = "camunda.secrets.";
 
-  /** Reads and writes the msgpack encoding of job variables as Jackson trees. */
-  private final ObjectMapper variablesMapper = new ObjectMapper(new MessagePackFactory());
-
-  private final SecretStoreRegistry secretStoreRegistry;
+  private final JobSecretValues secretValues;
 
   // accumulated per activation command while the collector checks and appends jobs, consumed
   // (and reset) by injectSecretValues; the collector's reset() bounds a new command
-  private final Map<SecretReference, String> values = new HashMap<>();
   private final List<JobWithCachedSecrets> jobsWithCachedSecrets = new ArrayList<>();
 
   // jobs skipped for a non-cached secret reference, grouped by reference so the processor can emit
@@ -89,41 +69,26 @@ public final class JobSecretInjector {
   private final Map<SecretReference, List<Long>> jobsWithNonCachedSecrets = new LinkedHashMap<>();
 
   public JobSecretInjector(final SecretStoreRegistry secretStoreRegistry) {
-    this.secretStoreRegistry = secretStoreRegistry;
+    secretValues = new JobSecretValues(secretStoreRegistry);
   }
 
   /** Discards any state accumulated for a previous activation command. */
   public void reset() {
-    values.clear();
+    secretValues.reset();
     jobsWithCachedSecrets.clear();
     jobsWithNonCachedSecrets.clear();
   }
 
   /**
-   * Checks each secret reference of the job (stored on the {@link JobRecord} at creation) for a
-   * value its store already holds locally, materializing the values and the job's secrets once. No
-   * store is read, so this cannot block the stream processor. Every reference is checked, even
-   * after the first miss, so the caller sees all non-cached references of the job. A failing lookup
-   * propagates to the caller and fails the activation command.
+   * Checks the job's secret references for cached values, see {@link
+   * JobSecretValues#check(JobRecord)}.
    *
    * <p>The collector skips a job with any non-cached reference and registers it via {@link
    * #registerForResolution}, so the processor can request the background resolution of the
    * non-cached references and park the job until it is resolved.
    */
   public SecretCheckResult checkSecrets(final JobRecord job) {
-    if (!job.hasSecretReferences()) {
-      return SecretCheckResult.NO_SECRETS;
-    }
-    final List<Secret> cachedSecrets = new ArrayList<>();
-    final List<Secret> nonCachedSecrets = new ArrayList<>();
-    for (final Secret secret : secretsOf(job)) {
-      if (resolveIntoValues(secret.reference())) {
-        cachedSecrets.add(secret);
-      } else {
-        nonCachedSecrets.add(secret);
-      }
-    }
-    return new SecretCheckResult(cachedSecrets, nonCachedSecrets);
+    return secretValues.check(job);
   }
 
   /**
@@ -176,24 +141,6 @@ public final class JobSecretInjector {
   }
 
   /**
-   * Resolves the reference into the materialized values, or returns {@code false} when no store
-   * holds a value for it locally. A reference resolved before is not looked up again. The lookup
-   * reads no store, so it cannot block the stream processor; a lookup failure is deliberately not
-   * caught here: it propagates and fails the activation command.
-   */
-  private boolean resolveIntoValues(final SecretReference reference) {
-    if (values.containsKey(reference)) {
-      return true;
-    }
-    // the store lookup is exact: a reference naming no configured store addresses none
-    final Optional<String> value =
-        Optional.ofNullable(secretStoreRegistry.getStores().get(reference.storeId()))
-            .flatMap(store -> store.lookupLocal(reference.name()));
-    value.ifPresent(cachedValue -> values.put(reference, cachedValue));
-    return value.isPresent();
-  }
-
-  /**
    * Injects the cached values into the secret placeholders of the registered jobs, replacing the
    * variables of the response batch in place. Consumes the values and jobs accumulated since the
    * last reset (the engine processor is single-threaded, so the collection of a command always
@@ -217,7 +164,9 @@ public final class JobSecretInjector {
       for (final JobWithCachedSecrets jobWithSecrets : jobsWithCachedSecrets) {
         final byte[] injected;
         try {
-          injected = injectedVariablesOf(jobWithSecrets);
+          injected =
+              secretValues.injectedVariablesOf(
+                  jobWithSecrets.job(), jobWithSecrets.cachedSecrets());
         } catch (final Exception e) {
           return Optional.of(
               dropJobsWhoseInjectionFailed(responseBatch, activatedBatch, jobWithSecrets, e));
@@ -242,23 +191,6 @@ public final class JobSecretInjector {
     } finally {
       reset();
     }
-  }
-
-  /**
-   * Returns the job's variables with every secret placeholder replaced by its cached value, or
-   * {@code null} when no placeholder was found. A failure propagates to the caller, which drops the
-   * job from the activation and reports it for an incident.
-   */
-  private byte[] injectedVariablesOf(final JobWithCachedSecrets jobWithSecrets) throws IOException {
-    final DirectBuffer variables = jobWithSecrets.job().getVariablesBuffer();
-    final JsonNode document = variablesMapper.readTree(new DirectBufferInputStream(variables));
-    // a placeholder can be absent (e.g. fetchVariables excluded the variable); an unchanged job
-    // must keep its original encoding and stay out of the growth budget, so it returns null
-    boolean changed = false;
-    for (final Secret secret : jobWithSecrets.cachedSecrets()) {
-      changed |= replaceInLeaf(document, secret, values.get(secret.reference()));
-    }
-    return changed ? variablesMapper.writeValueAsBytes(document) : null;
   }
 
   /**
@@ -334,60 +266,6 @@ public final class JobSecretInjector {
   }
 
   /**
-   * Replaces the placeholder in the text leaf addressed by the secret's JSON pointer. Pointers that
-   * do not address a text leaf of an object (e.g. the path no longer matches the variables, or it
-   * runs into an array) are skipped defensively.
-   */
-  private static boolean replaceInLeaf(
-      final JsonNode document, final Secret secret, final String value) {
-    if (!secret.path().startsWith("/")) {
-      return false;
-    }
-    final JsonPointer pointer = JsonPointer.compile(secret.path());
-    final JsonNode parent = document.at(pointer.head());
-    final JsonNode leaf = document.at(pointer);
-    if (!parent.isObject() || !leaf.isTextual()) {
-      return false;
-    }
-    final String text = leaf.textValue();
-    final String replaced = text.replace(secret.placeholder(), value);
-    if (replaced.equals(text)) {
-      return false;
-    }
-    ((ObjectNode) parent).put(pointer.last().getMatchingProperty(), replaced);
-    return true;
-  }
-
-  /**
-   * Materializes the job's secret references, longest placeholder first so a reference name that is
-   * a prefix of another (e.g. {@code token} vs {@code token2}) cannot corrupt the longer
-   * placeholder when both are injected into the same leaf.
-   */
-  private static List<Secret> secretsOf(final JobRecord job) {
-    final List<Secret> secrets = new ArrayList<>();
-    for (final JobSecretReference reference : job.secretReferences()) {
-      secrets.add(
-          new Secret(
-              new SecretReference(reference.getStoreId(), reference.getSecretReference()),
-              reference.getPath(),
-              SECRET_REFERENCE_PREFIX + reference.getSecretReference()));
-    }
-    if (secrets.size() > 1) {
-      secrets.sort(comparingInt((Secret secret) -> secret.placeholder().length()).reversed());
-    }
-    return secrets;
-  }
-
-  /**
-   * The result of {@link #checkSecrets} for one job: the job's secrets with a cached value and
-   * those without one, each materialized once (both empty for a job without secret references). A
-   * job with any non-cached secret must not be activated.
-   */
-  public record SecretCheckResult(List<Secret> cachedSecrets, List<Secret> nonCachedSecrets) {
-    static final SecretCheckResult NO_SECRETS = new SecretCheckResult(List.of(), List.of());
-  }
-
-  /**
    * A job dropped from the batch whose secret values can never fit: injecting them would grow the
    * batch by {@code growth} bytes, more than the whole budget any activation batch has to spare.
    */
@@ -395,8 +273,6 @@ public final class JobSecretInjector {
 
   /** A job dropped from the batch because injecting its secret values failed. */
   public record FailedInjectionJob(long jobKey, JobRecord job) implements DroppedJob {}
-
-  record Secret(SecretReference reference, String path, String placeholder) {}
 
   /** A registered job: its index in the batch, the job, and its cached secrets. */
   record JobWithCachedSecrets(int index, JobRecord job, List<Secret> cachedSecrets) {}
